PR review changes and survey styling update
