+++ v2/client/src/helpers/createSum.js
@@ -1,0 +1,1 @@
+export const sum = arr => arr.reduce((a, b) => a + b);
